ShareAdd.jl

This Julia package is intended for interactive use, and it's aim is to help you in reducing clutter in your main shared environment, and thus avoid package incompatibility problems. It exports two macros: @usingany and @usingtmp, envisioned for two different workflows. The package also provides several utility functions for managing shared environments.

@usingany macro

This macro makes package(s) available, if they are not already, and loads them with using keyword.

If a package is available in an environment in LOAD_PATH, that's OK.

If a package is available in a shared environment, this environment will be pushed into LOAD_PATH.

Otherwise if it can be installed, you will be prompted to select an environment to install the package(s).

If the package is not listed in any registry, an error will be thrown.

# simplest usage case
@usingany SomePackage

For more usage options see @usingany docs .

@usingany usage example

Let's assume, while working on your package MyPackage, we temporarily need packages TOML, Plots, and Chairmarks. However, they shouldn't be added permanently to your package dependencies. Furthermore, from the package BenchmarkTools we need only the macro @btime and the function save. We also need Unitful, which is already an installed dependence of MyPackage.

TOML is available in the stdlib, Plots and BenchmarkTools you already put into a shared environment @utilities, and Chairmarks is not on your computer yet.

Now, first, you add ShareAdd to your "main" (standard) enviroment, making it available at all times:

]
(YourEnv) pkg> activate 
  Activating project at `~/.julia/environments/v1.11`

(@v1.11) pkg> add ShareAdd
(...)
(@v1.11 pkg> activate . # back to your environment
(YourEnv) pkg> 

By that occasion you may also want to clean your standard environment: It is generally not recommended having a lot of packages there.

Now, the only thing you need is to type into REPL (or adding to your script) the following lines:

using ShareAdd
@usingany Unitful, TOML, Plots, Chairmarks
@usingany BenchmarkTools: @btime, save

As Chairmarks was not installed yet, you will be asked as to where to install it. You may e.g. add it to your existing @utilities shared environment, or let create a new environment @Chairmarks and put it there.

Afrerwards @utilities (and @Chairmarks, if created) will be added to LOAD_PATH, making their packages available.

Finally, the macros will execute using Unitful, TOML, Plots, Chairmarks resp. using BenchmarkTools: @btime, save - and that's it. Enjoy!

@usingany with updates

By setting the corresponding kwarg, it is possible to first update the packages and/or environments prior to execution of import. E.g. the following command would update the packages Pkg1, Pkg2 in their shared environments:

using ShareAdd
@usingany update_pkg = true Pkg1, Pkg2

@usingany without explicitly calling @usingany

ShareAdd.jl can be combined nicely with BasicAutoloads.jl. See this Discourse post to learn how get packages silently loaded if you call their functions in the REPL - e.g. if you type mean([1,2,3]) or 1.55u"V".

Versioned manifests

If currently used Julia version supports versioned manifests (i.e. >= v1.11), then on any updates using ShareAdd package (see ShareAdd.update), a versioned manifest will be created in each updated env. The function ShareAdd.make_current_mnf can also be used to create a versioned manifest in a specified environment without updating it.

@usingtmp macro

This macro activates a temporary environment, optionally installs packages into it, and loads them with using keyword.

If current environment is already a temporary one, environment is not changed.

If current env was a project (not package!), a temporary env will be activated.

If current env was a package (under development), e.g. MyPkg, a temporary env will be activated, AND MyPkg will be dev-ed in that temporary env.

Thу last one is actually the interesting case, as you can continue to work on MyPkg, while temporarily having additional packages available.

If @usingtmp was called with arguments, the corresponding packages will be installed into that temporary env, and imported with using keyword.

using ShareAdd
@usingtmp Foo, Bar
@usingtmp Baz: quux

Some other functions and usage cases

The functions ShareAdd.info(), ShareAdd.update(), ShareAdd.delete() do what their names say.

The function ShareAdd.make_importable also does what it says. It is used internally by @usingany, but it can also be used separately in special cases, e.g. if you need using A as B syntax, or want to import a package via import statement instead of using:

using ShareAdd
make_importable("Foo")
import Foo

Reference

Exported macros

ShareAdd.@usingany — Macro

@usingany pkg
@usingany pkg1, pkg2, ... 
@usingany pkg: fn
@usingany pkg: fn, @mcr, ... 
@usingany kwarg = true [pkg...]

Makes package(s) available, if they are not already, and loads them with using keyword.

If a package is available in an environment in LOAD_PATH, that's OK.

If a package is available in a shared environment, this environment will be pushed into LOAD_PATH.

Otherwise if package(s) can be installed, you will be prompted to select an environment to install each package.

If the package is not listed in any registry, an error will be thrown.

The macro can be called with keyword arguments:

update_pkg::Bool: if set to true, first updates the package(s) to be imported by the macro

update_env::Bool: first update the shared environments currently in the LOAD_PATH

update_all::Bool: first update the package to be imported in ALL shared environments where it is present

If Julia version supports versioned manifests, on any updates a versioned manifest will be created in each updated env. See also make_current_mnf and update.

If update_all or update_env kwarg is set, @usingany can be called without specifying any package(s) for import. If update_pkg kwarg is set, package(s) to import must be specified.

This macro is exported.

Examples

julia> @usingany Foo, Bar
julia> @usingany Baz: quux
julia> @usingany update_all = true
julia> @usingany update_pkg = true Qux

ShareAdd.@usingtmp — Macro

@usingtmp 
@usingtmp pkg
@usingtmp pkg1, pkg2, ... 
@usingtmp pkg: fn
@usingtmp pkg: fn, @mcr, ...

Activates a temporary environment, optionally installs packages into it and loads them with using keyword.

If current environment is a temporary one, environment is not changed.

If current env was a project (not package!), a temporary env will be activated.

If current env was a package (under development), e.g. MyPkg, a temporary env will be activated, AND MyPkg will be dev-ed in that temporary env.

Afterwards, if @usingtmp was called with arguments, the corresponding packages will be installed into that temporary env, and imported with using keyword.

This macro is exported.

Public types

ShareAdd.EnvInfo — Type

mutable struct EnvInfo
EnvInfo(name::AbstractString) -> EnvInfo

name::String - name of the environment

path::String - path of the environment's folder

pkgs::Vector{String} - list of packages in the environment

in_path::Bool - whether the environment is in LOAD_PATH

Examples

julia> ShareAdd.EnvInfo("@DocumenterTools")
ShareAdd.EnvInfo("DocumenterTools", "/Users/eben60/.julia/environments/DocumenterTools", Set(["DocumenterTools"]), false, false, true, false, false)

ShareAdd.PackageInfo — Type

mutable struct PackageInfo

name::String - name of the package

envs::Vector{EnvInfo} - list of environments in which the package is present

in_path::Bool - whether any of the environments is in LOAD_PATH

Public functions

ShareAdd.current_env — Method

current_env(; depot = first(DEPOT_PATH)) -> EnvInfo

Returns information about the current active environment as an EnvInfo object.

This function is public, not exported.

ShareAdd.delete — Method

delete(nms::Union{String, Vector{String}}; inall=false)

Deletes shared envs, or packages therein.

If the provided argument is name(s) of shared environment(s), as specified by leading "@" in the names(s): then deletes the shared environment(s) by erasing their directory.

Otherwise, the provided name(s) are package names: then for each package pkg deletes it from it's shared environment. Deletes this environment if it was the only package there. If the package may be present in multiple environments, and you want to delete it from all of them, set inall=true.

Keyword argument

inall=false: If set to true, would delete package from multiple environments. Has no effect, if provided nms is/are env name(s).

Returns nothing.

This function is public, not exported.

ShareAdd.info — Method

info(nms::Union{Nothing, String, Vector{String}} = nothing; 
    by_env=true, listing=nothing, std_lib=false, upgradable=false, disp_rslt=true, ret_rslt=false)

Prints out and/or returns information about shared environments.

Argument

nms: Name(s) of package(s) or environment(s) to return the information on. Environment names must start with "@". Package and env names cannot be provided together in one array.

Keyword arguments

by_env=true: whether to print out results as a Dict of pairs like @env => [pkg1, ...], or pkg => [@env1, ...]. Has no effect on returned (if any) results.

listing=nothing: this kwarg can be nothing, :envs, or :pkgs. If one of these two Symbols is provided, the result is printed as a vector of envs or pkgs, resp. In this case by_env is ignored. Has no effect on returned (if any) results

upgradable=false: if true, all other kwargs will be ignored, and only upgradable packages with installed vs. most recent versions will be printed by environment.

disp_rslt=true: whether to print out results.

ret_rslt=false: whether the function returns anything. If set to true, it returns a NamedTuple (; env_dict, pkg_dict, envs, pkgs, absent), where the two first elements are Dicts with keywords correspondingly by env or by pkg; envs and pkgs are vectors of respective elements, and absent are those names provided through the nms argument, which are not contained in the shared envs. Names of envs in the returned data are without leading "@".

This function is public, but not exported, as to avoid possible name conflicts.

# Examples

julia> ShareAdd.info(["BenchmarkTools", "Chairmarks"])
The following packages are not in any shared env:
    ["Chairmarks"]

Found pkgs/envs:
  @BenchmarkTools
   => ["BenchmarkTools"]
  @Tools
   => ["BenchmarkTools"]

julia> ShareAdd.info(["DataFrames", "CSV"]; by_env=false)
  CSV
   => ["@DataFrames"]
  DataFrames
   => ["@DataFrames"]

julia> ShareAdd.info("StaticArrays"; upgradable=true)
  @StaticArrays
    StaticArrays: 1.9.8 --> 1.9.10   

ShareAdd.make_current_mnf — Method

make_current_mnf(path_or_name) -> path
make_current_mnf(; current::Bool) -> path
make_current_mnf(env::EnvInfo) -> path

Creates a versioned manifest

If called make_current_mnf(; current=true), the current environment will be processed by this function.

path_or_name can name of a shared environment starting with @, or a path to any environment.

If currently executed Julia version doesn't support version-specific manifests, do nothing.

Else, if a versioned manifest for current Julia already exists, do nothing.

Else, if the environment is the main shared env for the current Julia version (e.g. "@v1.11" for Julia v1.11), do nothing.

Else, is a (versioned) manifest for an older Julia exists in the given directory, copy it to a file named according to the current Julia version, e.g. Manifest-v1.11.toml.

Else, create empty one.

Returns path to the created or existing manifest.

This function is public, not exported.

ShareAdd.make_importable — Method

make_importable(pkg::AbstractString)
make_importable(pkgs::AbstractVector{<:AbstractString})
make_importable(pkg1, pkg2, ...)
make_importable(::Nothing) => :success

Checks packages (by name only, UUIDs not supported!), prompts to install packages which are not in any shared environment, and adds relevant shared environments to LOAD_PATH.

make_importable is used internally by @usingany, but it can be used separately e.g. if you e.g. want to import a package via import statement instead of using.

Returns :success if the operation was successful, and nothing if the user selected "Quit. Do Nothing." on any of the prompts.

Throws an error on unavailable packages.

Examples

julia> using ShareAdd
julia> make_importable("Foo")
:success
julia> import Foo 

julia> using ShareAdd
julia> make_importable("Foo")
:success
julia> using Foo: bazaar as baz  # @usingany Foo: bazaar as baz is not a supported syntax

This function is public, not exported.

ShareAdd.reset — Method

reset()

Resets the LOAD_PATH to it's default value of ["@", "@v#.#", "@stdlib"], thus removing any manually added paths.

This function is public, not exported.

ShareAdd.sh_add — Method

sh_add(env_name::AbstractString; depot = first(DEPOT_PATH)) -> Vector{String}
sh_add(env_names::AbstractVector{<:AbstractString}; depot = first(DEPOT_PATH)) -> Vector{String}
sh_add(env_name::AbstractString, ARGS...; depot = first(DEPOT_PATH)) -> Vector{String}

Adds shared environment(s) to LOAD_PATH, making the corresponding packages all available in the current session.

Returns the list of all packages in the added environments as a Vector{String}.

Examples

julia> sh_add("@StatPackages")
3-element Vector{String}:
 "Arrow"
 "CSV"
 "DataFrames"

julia> sh_add(["@StatPackages", "@Makie"])
4-element Vector{String}:
 "Arrow"
 "CSV"
 "DataFrames"
 "Makie"

julia> sh_add("@StatPackages", "@Makie")
4-element Vector{String}:
 "Arrow"
 "CSV"
 "DataFrames"
 "Makie"

This function is public, not exported.

ShareAdd.update — Function

update()
update(nm::AbstractString)
update(nm::Vector{<:AbstractString})
update(env::AbstractString, pkgs::Union{AbstractString, Vector{<:AbstractString}}) 
update(env::EnvInfo, pkgs::Union{Nothing, S, Vector{S}} = Nothing) where S <: AbstractString

Called with no arguments, updates all shared environments.

Called with a single argument nm::String starting with "@", updates the shared environment nm.

Called with a single argument nm::String not starting with "@", updates the package nm in all shared environments.

Called with a single argument nm::Vector{String}, updates the packages and/or environments in nm.

Called with two arguments env and pkgs, updates the package(s) pkgs in the environment env.

If Julia version supports version-specific manifest, then on any updates a versioned manifest will be created in each updated env. See also make_current_mnf.

Returnes nothing.

This function is public, not exported.